+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Examples.hs
@@ -283,6 +283,7 @@ examplePoolDistr =
               1
               (CompactCoin 1)
               exampleVrfVerKeyHash
+              SNothing
           )
         ]
     )