+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/State/Query/Examples.hs
@@ -103,6 +103,7 @@ import Test.Cardano.Ledger.Conway.Examples (
  )
 import Test.Cardano.Ledger.Core.Rational (unsafeBoundRational)
 import Test.Cardano.Ledger.Core.Utils (mkDummySafeHash)
+import Test.Cardano.Ledger.Dijkstra.Examples (exampleBlsKey)
 import Test.Cardano.Ledger.Era (EraTest (..))
 import Test.Cardano.Ledger.Shelley.Examples (
   examplePoolDistr,
@@ -248,6 +249,7 @@ querySetSnapshotStakePoolDistrExamples =
                     { individualPoolStake = 1 % 4
                     , individualTotalPoolStake = CompactCoin 1_000_000_000
                     , individualPoolStakeVrf = exampleVrfVerKeyHash
+                    , individualPoolStakeBls = SNothing
                     }
                 )
               ,
@@ -256,6 +258,7 @@ querySetSnapshotStakePoolDistrExamples =
                     { individualPoolStake = 3 % 8
                     , individualTotalPoolStake = CompactCoin 5_000_000_000
                     , individualPoolStakeVrf = exampleVrfVerKeyHash
+                    , individualPoolStakeBls = SJust exampleBlsKey
                     }
                 )
               ]